requirements.txt
Django>=4.2,<5.0
mysql-connector==2.2.9
virtualenv==20.0.31
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
Pillow
requests
django-core
mysqlclient
# psycopg
psycopg2-binary
python-dotenv
pymysql
uvicorn